+++ CHANGELOG.md
@@ -17,3 +17,4 @@
 - Improved execution of action as daemon.
 - Force execution of request even if the LCP operation failed.
 - Accept list of strings for eBPF-Program code.
+- Added Event management.